requirements.txt
mido>=1.2.10
requests>=2.0
python-dotenv>=1.0

# MIDI backend
python-rtmidi>=1.4.0
pyusb>=1.2.0

# Testing
pytest>=7.0.0
requests-mock>=1.11.0
